+++ components/TreeFiles/Branch.tsx
@@ -1,6 +1,6 @@
 import React, {FC, useState} from 'react';
 import Node from './Node';
-import {DATAType} from "../../assets/constans";
+import {DATAType} from "../../types/types_data_from_server/dataType";
 
 interface BranchProps {
   item: DATAType
