+++ components/TreeFiles/Tree.tsx
@@ -1,6 +1,7 @@
 import React, {FC} from 'react';
 import Branch from "./Branch";
-import {DATAType} from "../../assets/constans";
+import {DATAType} from "../../types/types_data_from_server/dataType";
+
 
 interface TreeProps {
   data: DATAType[]
