+++ types/types_data_from_server/dataType.ts
@@ -1,0 +1,5 @@
+export type DATAType = {
+  id?: string
+  label?: string
+  children?: DATAType[]
+}
